[Chore] #46 - 코드 이동
