+++ Napzak-iOS/Napzak-iOS/Network/Product/ProductService.swift
@@ -8,11 +8,19 @@
 import Moya
 
 protocol ProductServiceProtocol {
-    
+    func postSellRegister(sellRegisterProduct: SellRegisterRequestDTO) async -> Result<SellRegisterResponseDTO, NetworkError>
+    func postBuyRegister(buyRegisterProduct: BuyRegisterRequestDTO) async -> Result<BuyRegisterResponseDTO, NetworkError>
 }
 
 final class ProductService: BaseService, ProductServiceProtocol {
    
-//    private let provider = MoyaProvider<ProductAPI>.init(plugins: [MoyaPlugin()])
+    private let provider = MoyaProvider<ProductAPI>.init(plugins: [MoyaPlugin()])
     
+    func postSellRegister(sellRegisterProduct: SellRegisterRequestDTO) async -> Result<SellRegisterResponseDTO, NetworkError> {
+        return await request(provider, .sellRegister(registerItem: sellRegisterProduct))
+    }
+
+    func postBuyRegister(buyRegisterProduct: BuyRegisterRequestDTO) async -> Result<BuyRegisterResponseDTO, NetworkError> {
+        return await request(provider, .buyRegister(registerItem: buyRegisterProduct))
+    }
 }
